Split the Decoder to account for wasm/emscripten (lack of) fs support
This mirrors what we do with the Encoder -- a "SimpleDecoder" for most
of the functionality (might update the naming :thinking:), and the
"Decoder" for all that + some fs helper methods.

Some slight excitement with the overload of `decode()` -- if that
becomes an issue we'll change the name in the Decoder class (to
`decode_File()` perhaps).

(extra context for the record: we could do FS support in wasm/emscripten
if we really wanted to. AFAIK, we don't.)
